Extract FatBuffer to its own module
